requirements.txt
alembic==1.4.3
blinker==1.4
click==7.1.2
decorator==4.4.2
dnspython==2.0.0
email-validator==1.1.2
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
infinity==1.5
intervals==0.9.0
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.20
SQLAlchemy-Utils==0.36.8
validators==0.18.1
Werkzeug==1.0.1
WTForms==2.3.3
WTForms-Alchemy==0.17.0
WTForms-Components==0.10.4
